Update WHMCS-Discord-Notifications.php
This code will add a hook for the ServiceRenewed event, and send a notification to the Discord channel if the $serviceRenewed notification setting is true. The $vars array passed to the hook function contains information about the renewed service, such as the serviceid and invoiceid. This information can be used to include details about the renewed service in the notification.